Just some minor updates - trying to get ANT & BLE speed working stably, but that's not happening.  ANT is stable, but wrong.  BLE is closer, but not stable.

+++ firmware/emotion/libraries/speed.h
@@ -45,7 +45,7 @@ typedef struct speed_event_s
 {
 	uint16_t	period_2048;															// Time period in 1/2048's of second since the prior event.
 	uint16_t 	event_time_2048;													// Time stamp of the associated wheel revolution in 1/2048's of a second.
-	uint8_t 	accum_wheel_revs;													// Count of accumulated wheel revolutions at this time stamp.
+	uint32_t 	accum_wheel_revs;													// Count of accumulated wheel revolutions at this time stamp.
 	uint32_t 	accum_flywheel_revs;											// Count of flywheel revs.
 	float			speed_mps;																// Speed in meters per second.	
 } speed_event_t;
